+++ src/pyfx/config/config_parser.py
@@ -1,59 +1,57 @@
-import importlib.resources
 import pathlib
 
 import dacite
 import yamale
 from first import first
+from loguru import logger
 from yamale import YamaleError
-from yamale.validators import DefaultValidators
 
-from ..error import PyfxException
-from .validators import Options
-from .config import Configuration
+from pyfx.error import PyfxException
+from pyfx.config import PyfxConfiguration
+from pyfx.config.yaml import yaml_path
+
+
+def load(data_file, data_class):
+    """Loads a configuration into a dataclass."""
+
+    try:
+        data = yamale.make_data(data_file.absolute())
+        # Result of `yamale#make_data` is composed as [(config, path)...]
+        return dacite.from_dict(data=data[0][0], data_class=data_class)
+    except Exception as e:
+        logger.opt(exception=True).error(e)
+        raise PyfxException(f"Load {data_file} into {data_class} failed with "
+                            "unknown error.", e)
 
 
 def parse(config_file=None):
+    """Parses and validates the configuration file."""
+
     try:
-        return ConfigurationParser().parse(config_file)
+        schema_file = yaml_path / "config_schema.yml"
+        schema = yamale.make_schema(schema_file.absolute())
+
+        config_file = first([
+            # user provided config file
+            pathlib.Path(config_file) if config_file is not None else None,
+            # ~/.config/pyfx/config.yml
+            pathlib.Path.home() / ".config" / "pyfx" / "config.yml",
+            # ./yaml/config.yml
+            yaml_path / "config.yml"
+        ], key=lambda p: p is not None and p.exists() and p.is_file())
+        config = yamale.make_data(config_file.absolute())
+
+        # Validates the configuration file
+        yamale.validate(schema, config)
+
+        # Result of `yamale#make_data` is composed as [(config, path)...]
+        return dacite.from_dict(data=config[0][0], data_class=PyfxConfiguration)
     except YamaleError as e:
-        # catch and raise a more user-friendly error
+        # Catches and raises an error with a more user-friendly error message
         message = '\n'.join(e.message.split('\n')[1:]).strip()
-        raise PyfxException(f"Configuration Error: {message}.")
-
-
-class ConfigurationParser:
-    __CONFIG_PATHS = [
-        pathlib.Path.home() / ".config" / "pyfx" / "config.yml",
-    ]
-
-    def __init__(self):
-        self._schema = self.__load_schema()
-
-    def parse(self, config_file=None):
-        config = self.__load_config(config_file)
-        yamale.validate(self._schema, config)
-        # config is composed as [(config, path)...]
-        return dacite.from_dict(data_class=Configuration, data=config[0][0])
-
-    @staticmethod
-    def __load_schema():
-        validators = DefaultValidators.copy()
-        validators[Options.tag] = Options
-
-        contents = importlib.resources.read_text("pyfx.config", "schema.yml")
-        return yamale.make_schema(validators=validators, content=contents)
-
-    # noinspection PyBroadException
-    @staticmethod
-    def __load_config(config_file):
-        if config_file is None:
-            config_file = first(
-                ConfigurationParser.__CONFIG_PATHS,
-                key=lambda path: path.exists()
-            )
-        if config_file is not None:
-            return yamale.make_data(path=config_file)
-
-        # use default config
-        contents = importlib.resources.read_text("pyfx.config", "config.yml")
-        return yamale.make_data(content=contents)
+        raise PyfxException(f"Parse and validate Pyfx configuration file "
+                            f"failed with error: {message}.")
+    except Exception as e:
+        logger.opt(exception=True).error(e)
+        raise PyfxException("Parse Pyfx configuration file failed with unkown "
+                            "error.", e)