+++ src/pyfx/model/model.py
@@ -6,7 +6,7 @@ from .autocomplete import autocomplete
 
 class Model:
     """
-    pyfx model entry point, which loads and processes JSON data.
+    Pyfx model entry point, which processes JSON data.
 
     Currently it manages the following actions:
      * parses JSONPath query and returns new data