+++ src/pyfx/view/components/help_popup.py
@@ -5,8 +5,8 @@ import urwid
 from overrides import overrides
 
 from pyfx.view.common import SelectableText
-from pyfx.view.keymapper import AbstractComponentKeyMapper
-from pyfx.view.keymapper import KeyDefinition
+from pyfx.view.components.abstract_component_keys import BaseComponentKeyMapper
+from pyfx.view.components.abstract_component_keys import KeyDefinition
 
 
 class HelpPopUpKeys(KeyDefinition, Enum):
@@ -20,7 +20,7 @@ class HelpPopUpKeys(KeyDefinition, Enum):
 
 
 @dataclass(frozen=True)
-class HelpPopUpKeyMapper(AbstractComponentKeyMapper):
+class HelpPopUpKeyMapper(BaseComponentKeyMapper):
     cursor_up: str = "up"
     cursor_down: str = "down"
     exit: str = "esc"