+++ src/pyfx/view/theme/__init__.py
@@ -1,2 +1,0 @@
-"""Components that defines color scheme for different components inside Pyfx."""
-from .theme_config import ThemeConfiguration